Feat: Implement cache emulation on matrix multiplcation

+++ cache_assignment/Cache/cache.cpp
@@ -10,36 +10,36 @@ using namespace std;
 class cache 
 {
     private:
-        int cache_line_length;
-        int cache_size;
-        int lines;
-        int element_size;
-        int associativity;
-        int set_length;
-        map<int,list<int>> lru_caches;
-        vector<vector<int>> cache_block;
+        long cache_line_length;
+        long cache_size;
+        long lines;
+        long element_size;
+        long associativity;
+        long set_length;
+        map<long,list<long>> lru_caches;
+        vector<vector<long>> cache_block;
         vector<bool> tags;
 
         void initialize_cache()
         {
             /*Initialises the cache*/
             lines = cache_size / cache_line_length;
-            int elements_in_line = cache_line_length / element_size;
+            long elements_in_line = cache_line_length / element_size;
             if(lines > associativity)set_length = lines/associativity;
             else set_length = 1;
-            cache_block.resize(lines, vector<int>(elements_in_line, 0));
+            cache_block.resize(lines, vector<long>(elements_in_line, 0));
             tags.resize(lines,0);
 
             // initalise pq for sets
-            for(int i = 0 ; i <associativity; i++)
+            for(long i = 0 ; i <associativity; i++)
             {
-                for(int j = 0 ; j < set_length; j++ )
+                for(long j = 0 ; j < set_length; j++ )
                 {
                     lru_caches[i].push_back(j);
                 }
             }
         }
-        void update_lru(int target_set, int line_idx)
+        void update_lru(long target_set, long line_idx)
         {
             /* updates the lru cache , getting the most recently used to the front*/
             for(auto it = lru_caches[target_set].begin(); it != lru_caches[target_set].end(); ++it)
@@ -51,10 +51,10 @@ class cache
                 }
             }
         }
-        int replace_lru(int target_set)
+        long replace_lru(long target_set)
         {
             /*choses the index to be replaced and performs the replacement mainitaing property of LRU*/
-            int replacement_line;
+            long replacement_line;
             if(!lru_caches[target_set].empty())
             {
                 replacement_line = lru_caches[target_set].back();
@@ -63,15 +63,15 @@ class cache
             }
             return replacement_line;
         }
-        void fetch_cache_line(int line,int addr)
+        void fetch_cache_line(long line,long addr)
         {
             /* fetch_cache_lines and places the address and neighbours at the line*/
-            for(int j = 0 ; j < cache_block[line].size();j++)
+            for(long j = 0 ; j < cache_block[line].size();j++)
             {
                 cache_block[line][j] = addr+j;
             }
         }
-        int get_set(int addr)
+        long get_set(long addr)
         {
             /*Gives the set a particular address belongs*/
             if (addr>cache_size) addr = addr%cache_size;
@@ -79,23 +79,23 @@ class cache
         }
 
     public:
-        cache(int size, int line_length, int elem_size, int assoc) 
+        cache(long size, long line_length, long elem_size, long assoc) 
             : cache_size(size), cache_line_length(line_length), 
             element_size(elem_size), associativity(assoc)
         {
             initialize_cache();
         }
-        bool find(int addr)
+        bool find(long addr)
         {
             /*Finds address in cache, returns true for a hit and false for miss*/
-            int target_set = get_set(addr);
-            for(int i = target_set*set_length ; i < (target_set+1)*set_length;i++)
+            long target_set = get_set(addr);
+            for(long i = target_set*set_length ; i < (target_set+1)*set_length;i++)
             {
-                for(int j = 0 ; j < cache_block[i].size();j++)
+                for(long j = 0 ; j < cache_block[i].size();j++)
                 {
                     if(cache_block[i][j] == addr)
                     {
-                        int update_idx = i/set_length;
+                        long update_idx = i/set_length;
                         update_lru(target_set,update_idx);
                         if(tags[i] ==0)
                         {
@@ -107,8 +107,8 @@ class cache
                     }
                 }
             }
-            int replacement_idx = replace_lru(target_set);
-            int replacement_line  = (target_set*set_length) + replacement_idx;
+            long replacement_idx = replace_lru(target_set);
+            long replacement_line  = (target_set*set_length) + replacement_idx;
             tags[replacement_line]=1;
             fetch_cache_line(replacement_line,addr);
             return false;
@@ -116,9 +116,9 @@ class cache
         void display_cache()
         {
             /*displays the cache*/
-            for(int i = 0 ; i < cache_block.size(); i++)
+            for(long i = 0 ; i < cache_block.size(); i++)
             {
-                for(int j = 0; j <cache_block[i].size();j++)
+                for(long j = 0; j <cache_block[i].size();j++)
                 {
                     cout<<cache_block[i][j]<< " ";
                 }
